==> 1_pesquisa_e_desenvolvimento/1_Divisao_dos_dados.ipynb <==
# Divisão dos dados

Neste notebook é feita a divisão dos dados em dois conjuntos: um conjunto usado par treinar o modelo (**dados_treinamento.csv**) e outro usado para testar o modelo treinado e a implementação do modelo como um pacote (**dados_teste.csv**).

Os dataset original (**bodyPerformance.csv**) foi utlizado somente neste notebook. Nas proximas etapas sarão utilizados os datasets resultantes da sua divisão.

Fonte dos dados originais:
https://www.kaggle.com/datasets/kukuroo3/body-performance-data


import pandas as pd
from sklearn.model_selection import train_test_split

# leitura dos dados
df = pd.read_csv(".//dados//dados_originais//bodyPerformance.csv" ,sep=',');

# verificação
df.head(4)

# divide os dados
dados_train, dados_test = train_test_split(df,
                                           test_size=0.1,
                                           random_state=0)

# verificação
print("Dimensão dos dados de treinamento: ",dados_train.shape)
print("Dimensão dos dados de teste: ",dados_test.shape)

# salva os dados
dados_train.to_csv('.//dados//dados_originais//dados_treinamento.csv', index=False)
dados_test.to_csv('.//dados//dados_originais//dados_teste.csv', index=False)
