==> 1_pesquisa_e_desenvolvimento/3_treinamento_do_modelo.ipynb <==
# Treinamento do modelos

Nesta seção feitos alguns treinamentos de modelos visando obter o que fornace o melhor resultado. Considerei 5 modelos para o estudo () cada modelo foi treinado considerando com as seguintes tranformações nos dados:

Etapas:

   **Treinamento manual de um modelo simples**: nesta etapa mostro o treinamentos manuais de modelo sem nenhuama transformação, o objjetivo desta etapa é entender os modelos e estabelecer valores limites para um optimização de hiperparâmetros
    
   **Otimização de hiperparâmetros:** raliza a optimização  de hiperparâmetros para os modelos. Nesta eta não são consideradas transformações nos dados.
   
   **Treinamento de modelo com transformações nos dados:**  Utilizando os hiperparametros obtidos na etapa anterior, treinei os modelos com algumas transformações no dados de treinamento:



Sequência para os treinamentos de modelo:

- Divide os dados
- (Faz uma transformação) *somente na etapa final
- Treina o modelo
- Testa o modelo

import pandas as pd
import matplotlib.pyplot as plt
import numpy as np

# transfomações
from sklearn.preprocessing import StandardScaler
from sklearn.preprocessing import QuantileTransformer
from sklearn.preprocessing import PowerTransformer
from sklearn.preprocessing import Normalizer

# para divisao dos dados
from sklearn.model_selection import train_test_split
from sklearn.model_selection import GridSearchCV
from sklearn import metrics

# modelos
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.neighbors import KNeighborsClassifier
from sklearn import svm

# para não mostrar mensagens com warnings
import warnings

import time

#leitura dos dados
df = pd.read_csv(".//dados//dados_originais//dados_treinamento.csv" ,sep=',');
df.head(3)

#Mapeia a variável 'gender' para valores inteiros
gender_map = {'M': 0, 'F': 1}
var = 'gender'
df[var] = df[var].map(gender_map)

# seleciona variaveis de treinamento e alvo
var_train=[var for var in df.columns if var != 'class'] 

X = df[var_train]   
y = df['class'] 

# divide os dados
X_train, X_test, y_train, y_test = train_test_split(X,y,
                                                    test_size=0.1,
                                                    random_state=0)

# Otimização de hiperparâmetros

## decision tree

# set up the model
tree = DecisionTreeClassifier(criterion="entropy",
                             random_state=0,
                             min_impurity_decrease=0
                             )

# determine the hyperparameter space
param_grid = dict(
    max_depth=[8,10,12,14,16],
    min_samples_split=[8,12,14,16,18,20,22,24],
    min_samples_leaf=[3,5,7],    
    )

# set up the search
search = GridSearchCV(tree, param_grid, scoring='accuracy',
                      cv=3, refit=True)

# find best hyperparameters
search.fit(X_train, y_train)

# Seleciona resultados desejados

df_results = pd.DataFrame(search.cv_results_)

columns=['mean_test_score','std_test_score', 'param_max_depth',
     'param_min_samples_leaf','param_min_samples_split', 
     'split0_test_score','split1_test_score','split2_test_score']
 
colunms_rename=['mean_score','std_score', 'max_depth',
     'min_samples_leaf','min_samples_split', 
     'split0_score','split1_score','split2_score']
    

df_results = df_results[columns]
df_results.columns=colunms_rename

# ordena os resultados
df_results.sort_values(by='mean_score',
                       ascending=False,
                       inplace=True)

df_results.reset_index(inplace=True, drop=True)
df_results.head(5)

# seleciona os melhores resultados
df_results_sel=df_results[df_results['mean_score']> 0.668]

# organiza os resultado para 'std_score' crescente
df_results_sel.sort_values(by='std_score',
                           ascending=True,
                           inplace=True)

df_results_sel.reset_index(inplace=True, drop=True)
df_results_sel.head(3)

## Random forest

# set up the model
forest= RandomForestClassifier(bootstrap=True,
                               max_samples=None,
                               random_state=0)

# determine the hyperparameter space
param_grid = dict(
    max_depth=[12,14,16,18,20],
    max_features=[5,6,8],
    n_estimators=[40,50,60,80],    
    )

# set up the search
search = GridSearchCV(forest, param_grid, scoring='accuracy',
                      cv=3, refit=True)

# find best hyperparameters
search.fit(X_train, y_train)

# Seleciona resultados desejados

df_results = pd.DataFrame(search.cv_results_)

columns=['mean_test_score', 'std_test_score',  'param_max_depth',
         'param_max_features', 'param_n_estimators', 
         'split0_test_score', 'split1_test_score', 'split2_test_score']
 
colunms_rename=['mean_score', 'std_score',  'max_depth',
             'max_features', 'n_estimators',
             'split0_score', 'split1_score', 'split2_score']

df_results = df_results[columns]
df_results.columns=colunms_rename

# ordena os resultados
df_results.sort_values(by='mean_score',
                       ascending=False,
                       inplace=True)

df_results.reset_index(inplace=True, drop=True)
df_results.head(5)

# seleciona os melhores resultados
df_results_sel=df_results[df_results['mean_score']> 0.732]

# organiza os resultado para 'std_score' crescente
df_results_sel.sort_values(by='std_score',
                           ascending=True,
                           inplace=True)

df_results_sel.reset_index(inplace=True, drop=True)
df_results_sel.head(3)

## Modelos com as variações

transformacoes={
    'sem_transformacao': [],     #dados sem tranformação
    'yeo_johnson': [PowerTransformer(method="yeo-johnson")],
    'quantile': [QuantileTransformer(output_distribution="normal")],
    'standart': [StandardScaler()],
    'normalization': [Normalizer()],
}

classificadores={'decisionTree':DecisionTreeClassifier(
                               criterion="entropy",
                               max_depth=10,
                               min_samples_leaf=5,
                               min_samples_split=18,                               
                               min_impurity_decrease=0,
                               random_state=0),
                
                'randomForest':RandomForestClassifier(                             
                             max_depth=14,
                             max_features=8,
                             n_estimators=80,
                             bootstrap=True,
                             max_samples=None,
                             random_state=0),
                 
                 'kNearestNeighbour':KNeighborsClassifier(
                     n_neighbors=50,
                     weights='distance'
                     ),
                                                          
                 'rbf_SVM':svm.SVC(kernel="rbf",random_state=0,
                        gamma=2.9,                
                        max_iter=1000,
                        C=0.1),
                                                          
                 'poly3_SVM':svm.SVC(kernel="poly",
                                     degree=5,
                                     random_state=0,
                                     gamma=2.9,                
                                     max_iter=5000,
                                     C=0.5)
                
                }
                

#colunas que serão transformadas
listaVar=['age', 'height_cm', 'weight_kg',
 'body fat_%', 'diastolic', 'systolic',
 'gripForce', 'sit and bend forward_cm',
 'sit-ups counts', 'broad jump_cm']

warnings.filterwarnings("ignore")

dict_metricas={}

for key in transformacoes.keys():
    
    print("--- Treinamento para dados com a transformação: "+key)
    t_ini=time.time()
    
    # copia os dados
    X_train_temp = X_train.copy()
    X_test_temp = X_test.copy()     
        
        
    # transforma os dados de treinamento e de teste
    if transformacoes[key]:        
        dados_train = X_train_temp[listaVar].values  
        dados_test = X_test_temp[listaVar].values
        
        obj = transformacoes[key][0].fit(dados_train)  
        
        X_train_temp[listaVar] = obj.transform(dados_train)  
        X_test_temp[listaVar] = obj.transform(dados_test)  


    #treina e testa os modelos
    metricas_modelo=[]
    for nome_modelo, clf in classificadores.items():    
        
        clf.fit(X_train_temp,y_train)   
        
        y_hat = clf.predict(X_test_temp)       
        metricas_modelo= metricas_modelo + [metrics.accuracy_score(y_test, y_hat)]
        
        print(nome_modelo+" treinado!")
        
    
    
    # armazena os resultados
    dict_metricas[key]=metricas_modelo    
    print("(tempo: "+str(time.time()-t_ini)+" seg.)\n")    

    

# apresentação dos resultados
df=pd.DataFrame(dict_metricas)
df.index=list(classificadores.keys())
df

#=====================================
# Nao rodar e nao apagar
#=====================================

# resultado anterior- sem escolha por optimização

# apresentação dos resultados
df=pd.DataFrame(dict_metricas)
df.index=list(classificadores.keys())
df
